+++ extensions/kill-process/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Kill Process Changelog
 
+## [UX Improvements] - 2026-02-11
+
+- Removed the redundant **Default Sort By** preference (sorting is now persisted via the command dropdown)
+
 ## [UX Improvements] - 2026-02-10
 
 - Reorganized preferences into clearer sections